fix: CodeRabbit review round on the routing contract PR
- Critical: VM send editors no longer index getSends() with an
  unchecked findSendIndex sentinel — tryGetEngineRoute returns false
  when the engine no longer holds the send (undo/redo/rejected-command
  divergence), and all four editors return early.
- refreshLocalSendId matches the local entry by sendId 0 instead of
  assuming append-at-end, and pops the phantom local send when a
  rejected command left the mirror ahead of the engine.
- EditSendCommand requests a graph rebuild for mute/sidechainOnly/
  postFader changes too (they alter compiled topology + PDC edges),
  not just destination changes.
- UIRoutingMap m_onEditSendLevel stored callback type widened to
  uint64_t (was int, narrowing the sendId).
- Muse identityPolicy declares sends stable_id; send ids are exposed
  as exact decimal strings (not doubles) in routing-graph and
  latency-report evidence.
- Serializer saves sendId as an exact decimal string and loads strings
  exactly, accepting legacy numeric ids only inside the exact
  JSON-integer range.
- Tests: wide send id (2^53+1) round-trips through serializer and Muse
  diagnostics; RoutingRenderSemanticsTest uses the portable temp
  directory pattern and bounds-checks the graph index lookup.
  238/238 ctest.

+++ AestraUI/Widgets/UIRoutingMap.h
@@ -285,7 +285,7 @@ private:
     std::function<void(uint32_t)> m_onNodeMuteToggle;
     std::function<void(uint32_t)> m_onNodeSoloToggle;
     std::function<void(uint32_t, uint64_t)> m_onRemoveSend;
-    std::function<void(uint32_t, int, float)> m_onEditSendLevel;
+    std::function<void(uint32_t, uint64_t, float)> m_onEditSendLevel;
 };
 
 } // namespace AestraUI
